Event: move type to a separate file, `event_error.go`
The reason for the filename will be apparent in a moment.
